CPR-710 move needs attention checks out of reclusterService
